+++ example/exampleConfig.ts
@@ -86,6 +86,7 @@ const zamzaConfig = {
     jobs: {
         cleanUpDeleteTimeoutMs: 60000,
         topicConfigPollingMs: 32000,
+        metadataFetcherMs: 12 * 62000,
     },
 };