+++ lib/db/models/index.ts
@@ -1,2 +1,4 @@
 export * from "./KeyIndexModel";
 export * from "./TopicConfigModel";
+export * from "./TopicMetadata";
+export * from "./LockModel";